// frontend/src/plugins/metal/preparations/ors/models/SamplePreparationOrs.ts
export interface SamplePreparationOrsStep {
    name: string;
    value1?: string;
    unit1?: string;
    value2?: string;
    unit2?: string;
    value3?: string;
    unit3?: string;
    logBookID?: string;
}

export interface SamplePreparationOrs {
    id: number;
    label: string;
    steps: SamplePreparationOrsStep[];
}